// src/types/api.ts
export interface definitions {
    LoginUser: {
        email: string;
        /** Format: password */
        password: string;
    };
    LoginUserRequest: {
        user: definitions["LoginUser"];
    };
    NewUser: {
        username: string;
        email: string;
        /** Format: password */
        password: string;
    };
    NewUserRequest: {
        user: definitions["NewUser"];
    };
    User: {
        email: string;
        token: string;
        username: string;
        bio: string;
        image: string;
    };
    UserResponse: {
        user: definitions["User"];
    };
    UpdateUser: {
        email?: string;
        token?: string;
        username?: string;
        bio?: string;
        image?: string;
    };
    UpdateUserRequest: {
        user: definitions["UpdateUser"];
    };
    ProfileResponse: {
        profile: definitions["Profile"];
    };
    Profile: {
        username: string;
        bio: string;
        image: string;
        following: boolean;
    };
    Article: {
        slug: string;
        title: string;
        description: string;
        body: string;
        tagList: string[];
        /** Format: date-time */
        createdAt: string;
        /** Format: date-time */
        updatedAt: string;
        favorited: boolean;
        favoritesCount: number;
        author: definitions["Profile"];
    };
    SingleArticleResponse: {
        article: definitions["Article"];
    };
    MultipleArticlesResponse: {
        articles: definitions["Article"][];
        articlesCount: number;
    };
    NewArticle: {
        title: string;
        description: string;
        body: string;
        tagList?: string[];
    };
    NewArticleRequest: {
        article: definitions["NewArticle"];
    };
    UpdateArticle: {
        title?: string;
        description?: string;
        body?: string;
    };
    UpdateArticleRequest: {
        article: definitions["UpdateArticle"];
    };
    Comment: {
        id: number;
        /** Format: date-time */
        createdAt: string;
        /** Format: date-time */
        updatedAt: string;
        body: string;
        author: definitions["Profile"];
    };
    SingleCommentResponse: {
        comment: definitions["Comment"];
    };
    MultipleCommentsResponse: {
        comments: definitions["Comment"][];
    };
    NewComment: {
        body: string;
    };
    NewCommentRequest: {
        comment: definitions["NewComment"];
    };
    TagsResponse: {
        tags: string[];
    };
    GenericErrorModel: {
        errors: {
            body: string[];
        };
    };
}